Optimize profile page with server-side data prefetching
- Split profile page into server component (page.tsx) and client component (ProfileClient.tsx)
- Prefetch getAllCards() and getNews() on server instead of client
- Reduces client-side API calls from 6 to 4
- Cards and news data now available instantly on page load
